Refactor exceptions into own namespace
